requirements.txt
Flask>=2.0
torch>=1.12.0
torchvision>=0.13.0
Pillow>=8.0.0
numpy>=1.21.0
opencv-python>=4.5.0
matplotlib>=3.4.0
```// filepath: c:\Users\Gautham\MiniProject\requirements.txt
Flask>=2.0
torch>=1.12.0
torchvision>=0.13.0
Pillow>=8.0.0
numpy>=1.21.0
opencv-python>=4.5.0
matplotlib>=3.4.0